requirements.txt
fastapi
uvicorn
websockets
groq
deepgram-sdk
elevenlabs
python-dotenv
aiohttp